test(create-webstone-app): fix import paths for refactored functions
